jetson/model_server_requirements.txt
flask==2.3.3
gunicorn==21.2.0
numpy==2.2.6
scikit-learn==1.6.1
pillow==10.0.0
requests==2.31.0